types: make arcane_zenith strict-clean and remove it from the mypy ledger
Fixes the single no-any-return: the sigmoid range-mapping helper built its
interpolant from np.exp (Any), so the max/min clamp returned Any against a
declared float; the interpolant is now wrapped to float at the source.

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)

+++ src/sc_neurocore/arcane_zenith.py
@@ -63,7 +63,7 @@ class ArcaneZenithCognitiveCore:
 
     def _map_to_range(self, w: float, min_val: float, max_val: float) -> float:
         """Smooth sigmoid mapping centered at 0.5 to prevent edge explosions."""
-        t = 1.0 / (1.0 + np.exp(-10.0 * (w - 0.5)))
+        t = float(1.0 / (1.0 + np.exp(-10.0 * (w - 0.5))))
         return max(min_val, min(max_val, min_val + t * (max_val - min_val)))
 
     def step(self, current: float) -> int:
